COLOR DETECTION
LabVIEW is a graphical programming language widely used for Automation of large scale projects. The Vision Assistant Module is integrated to acquire and process Images for Image Processing and Computer Vision applications. This project includes creating a color template using a test image accessed from a data set and then using the template, more blobs of the same color is identified and counted. The Region of Interest is used to concentrate the processing to a much lesser area and thereby increasing the efficiency  of the algorithm.